Harden log worker against exceptions. Also refactored it so that it's only defined in one place
